requirements.txt
absl-py==0.7.1
altgraph==0.16.1
aniso8601==6.0.0
argcomplete==1.8.2
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.3
beautifulsoup4==4.5.3
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Cython==0.29.6
decorator==4.4.0
docx2txt==0.6
EbookLib==0.15
Flask==1.0.2
flask-restplus==0.12.1
future==0.17.1
gast==0.2.2
gcloud==0.17.0
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-vision==0.36.0
google-pasta==0.1.4
googleapis-common-protos==1.5.9
grpcio==1.19.0
h5py==2.9.0
httplib2==0.12.1
idna==2.8
imageio==2.5.0
imutils==0.5.2
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
jws==0.1.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.3
macholib==1.11
mahotas==1.4.5
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
networkx==2.2
nguyendl==1.0
numpy==1.16.2
oauth2client==3.0.0
opencv-python==4.0.0.21
pandas==0.24.2
pdf2image==1.5.1
pdfminer.six==20181108
pefile==2018.8.8
pep8==1.7.1
Pillow==5.4.1
pipenv==2018.11.26
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycryptodome==3.4.3
PyInstaller==3.4
pylint==2.3.1
pymongo==3.7.2
PyMuPDF==1.14.12
pyparsing==2.3.1
Pyrebase==3.0.27
pyrsistent==0.14.11
pytesseract==0.2.6
python-dateutil==2.8.0
python-jwt==2.0.1
python-pptx==0.6.5
pytz==2019.1
PyWavelets==1.0.2
pywin32-ctypes==0.2.0
PyYAML==5.1
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
scikit-image==0.15.0
scipy==1.2.1
six==1.10.0
sortedcontainers==2.1.0
SpeechRecognition==3.6.3
SPIEPy==0.2.0
StringGenerator==0.3.1
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
typed-ast==1.3.1
urllib3==1.24.3
virtualenv==16.4.3
virtualenv-clone==0.5.2
waitress==1.2.1
Wand==0.5.2
Werkzeug==0.15.1
wrapt==1.11.1
xlrd==1.0.0
XlsxWriter==1.1.5
